requirements.txt
Django==2.2.12
django-environ==0.4.5
django-multiselectfield==0.1.12
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
